+++ apps/gizual-app/src/pages/main/main.tsx
@@ -1,13 +1,11 @@
 import { Container, Languages, parseLanguages } from "@app/charts";
+import { useMainController } from "@app/controllers";
+import { Canvas, SearchBar, TitleBar } from "@app/primitives";
 import { useWindowSize } from "@app/utils";
 import { observer } from "mobx-react-lite";
 import React from "react";
 import ReactGridLayout from "react-grid-layout";
 
-import { useMainController } from "../../controllers";
-import { TitleBar } from "../../primitives";
-import { Canvas } from "../../primitives/canvas";
-import { SearchBar } from "../../primitives/search-bar/search-bar";
 import { SettingsPage } from "../settings";
 
 import style from "./main.module.scss";