+++ apps/gizual-app/src/primitives/editor/editor.tsx
@@ -1,8 +1,8 @@
+import { useMainController } from "@app/controllers";
 import { observer } from "mobx-react-lite";
 import React from "react";
 
-import { useMainController } from "../../controllers";
-import { FileViewModel } from "../file/file.vm";
+import { FileViewModel } from "../file";
 
 import style from "./editor.module.scss";
 import { EditorViewModel } from "./editor.vm";