+++ apps/gizual-app/src/primitives/file/index.ts
@@ -1,1 +1,2 @@
 export { File } from "./file";
+export * from "./file.vm";