+++ apps/gizual-app/src/primitives/search-bar/search-bar.tsx
@@ -1,5 +1,6 @@
 /// <reference types="vite-plugin-svgr/client" />
 
+import { IconGitBranchLine, IconSearch } from "@app/assets";
 import { useMainController } from "@app/controllers";
 import { StreamLanguage } from "@codemirror/language";
 import { simpleMode } from "@codemirror/legacy-modes/mode/simple-mode";
@@ -13,8 +14,6 @@ import { observer } from "mobx-react-lite";
 import React from "react";
 import { createPortal } from "react-dom";
 
-import { ReactComponent as GitBranchLine } from "../../assets/icons/git-branch-line.svg";
-import { ReactComponent as SearchIcon } from "../../assets/icons/search.svg";
 import sharedStyle from "../css/shared-styles.module.scss";
 import { IconButton } from "../icon-button";
 import { Select } from "../select";
@@ -103,7 +102,7 @@ export const SearchBar = observer(({ vm: externalVm }: SearchBarProps) => {
           options={mainController.branchNames.map((b) => {
             return { label: b, value: b };
           })}
-          icon={<GitBranchLine />}
+          icon={<IconGitBranchLine />}
         />
         <div id="inputWrapper" className={style.SearchInputWrapper}>
           <SearchInput vm={vm} />
@@ -137,14 +136,14 @@ const SearchInput = observer(({ vm }: Required<SearchBarProps>) => {
       },
       {
         key: "Tab",
-        run: (view) => {
+        run: (_view) => {
           console.log("Tab");
           return true;
         },
       },
       {
         key: "ArrowDown",
-        run: (view) => {
+        run: (_view) => {
           console.log("ArrowDown");
           return true;
         },
@@ -229,7 +228,7 @@ const SearchInput = observer(({ vm }: Required<SearchBarProps>) => {
         onClick={() => vm.search()}
         aria-label="Search"
       >
-        <SearchIcon />
+        <IconSearch />
       </IconButton>
     </div>
   );