+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 
 - **app:canvas:** implement SVG export
-- **app:canvas:** implement full width visualization style 
+- **app:canvas:** implement full width visualization style
 
 ### Changes